Move most of the display code into the Inventory page Move the hints files one level down into device
