+++ resources/views/device/hints/Readme.hints.md
