+++ resources/views/device/hints/ubiquiti.hints
