--- resources/views/device/tabs/ports/hints/cisco.hints
@@ -1,1 +1,0 @@
-ciscoASR92024SZM:[{"0":[[{"Gi0\/0\/1":"sfp"},{"Gi0\/0\/0":"sfp"}],[{"Gi0\/0\/3":"sfp"},{"Gi0\/0\/2":"sfp"}],[{"Gi0\/0\/5":"rj45"},{"Gi0\/0\/4":"rj45"}],[{"Gi0\/0\/7":"rj45"},{"Gi0\/0\/6":"rj45"}],[{"Gi0\/0\/9":"rj45"},{"Gi0\/0\/8":"rj45"}],[{"Gi0\/0\/11":"rj45"},{"Gi0\/0\/10":"rj45"}],[{"Gi0\/0\/13":"rj45"},{"Gi0\/0\/12":"rj45"}],[{"Gi0\/0\/15":"rj45"},{"Gi0\/0\/14":"rj45"}],[{"Gi0\/0\/17":"rj45"},{"Gi0\/0\/16":"rj45"}],[{"Gi0\/0\/19":"rj45"},{"Gi0\/0\/18":"rj45"}],[{"Gi0\/0\/21":"rj45"},{"Gi0\/0\/20":"rj45"}],[{"Gi0\/0\/23":"rj45"},{"Gi0\/0\/22":"rj45"}]],"1":{"12":[{"Te0\/0\/25":"rj45"},{"Te0\/0\/24":"rj45"}],"13":[{"Te0\/0\/27":"sfp"},{"Te0\/0\/26":"sfp"}]},"3":{"28":[{"Gi0":"sfp"}]}}]
